+++ core-bundle/src/Fragment/Annotation/ContentElement.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Fragment\Annotations;
+namespace Contao\CoreBundle\Fragment\Annotation;
 
 use Doctrine\Common\Annotations\Annotation\Attribute;
 use Doctrine\Common\Annotations\Annotation\Attributes;
@@ -20,7 +20,7 @@ use Doctrine\Common\Annotations\Annotation\Target;
  * Annotation that can be used to define controller as a content element.
  *
  * @Annotation
- * @Target("CLASS", "METHOD")
+ * @Target({"CLASS", "METHOD"})
  * @Attributes({
  *     @Attribute("category", required = true, type = "string"),
  *     @Attribute("options", type = "array"),
@@ -30,50 +30,6 @@ use Doctrine\Common\Annotations\Annotation\Target;
  *     @Attribute("type", type = "string"),
  * })
  */
-final class ContentElement
+final class ContentElement extends Base
 {
-    /**
-     * @var string
-     */
-    public $category;
-
-    /**
-     * @var array
-     */
-    public $options = [];
-
-    /**
-     * @var string
-     */
-    public $renderer;
-
-    /**
-     * @var string
-     */
-    public $service;
-
-    /**
-     * @var string
-     */
-    public $template;
-
-    /**
-     * @var string
-     */
-    public $type;
-
-    /**
-     * Annotation constructor.
-     *
-     * @param array $values
-     */
-    public function __construct(array $values)
-    {
-        $this->category = $values['category'];
-        $this->options = (array) $values['options'];
-        $this->service = $values['service'];
-        $this->renderer = $values['renderer'];
-        $this->template = $values['template'];
-        $this->type = $values['type'];
-    }
 }
